requirements.txt
python-telegram-bot==20.8
uvicorn==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
openai==1.6.1
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.110.0
gunicorn
flask
python-telegram-bot
sqlalchemy
python-dotenv